Calculus Time
I started putting in some of the derivatives in, but it is far from
done. I have actually been working on this for a while, I just never
got around to committing since I have not finished my derivatives
method.
I also optimized function generation a bit and corrected some
parenthetical stuff for subtraction and division.