Completeish
After many unexpected bugs and fixes, I finally have a mostly functional
program. I wrote out the interface, cleaned up some of the
simplification, and perfected the ability to read functions. It still
has small bugs and issues, but it is largely usable.
As it is, the program prompts you to pick a difficulty and then gives
you 5 different functions which you must antidifferentiate, tweaking the
difficulty to fit your needs as you go. In the future, I plan to have it
give you a score and possibly time you, but this is perfect for now.